Go rid of Dash MATCH on DropdownAIO. But The table(s) now need to be initialised with a *page_size* that is the largest that will ever be displayed
NOTES:

Noticed that when I click the *table size* setting there's loads of dash component traffic with
the server. I can see now this is a problem with Dash MATCH. I use my *DropdownAIO* in
the *table size* control, BUT it's also used in each table row as the action dropdown, in both
my test table pages. The means that there can be over 200 instances of *DropdownAIO* in the TableAIO
examples! When one is selected they **ALL kick off** with network traffic.